made them all compilling))

+++ ft_strncat.c
@@ -23,7 +23,7 @@ char	*ft_strncat(char *dest, const char *src, size_t n)
 	while (dest[i])
 		i++;
 	j = 0;
-	while (src[j] && j < n)
+	while (src[j] && j < (int) n)
 	{
 		dest[i + j] = src[j];
 		j++;
